# Faster contact reading.