Cache column indices for cursor reading. This increases reading speed by about 10%.

For my contact list with 336 entries the measured timings are:

| | "old" loading | with changes |
|--|---|---|
| 1 | 249ms | 229ms |
| 2 | 265ms | 224ms |
| 3 | 282ms | 225ms |
| 4 | 281ms | 218ms |
| 5 | 256ms | 223ms |
| 6 | 272ms | 229ms |
| avg | 267.5ms | 224.6ms |

The changes are quite minor, so including this should not have negative side effect.